requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
mlxtend>=0.23
matplotlib>=3.7
seaborn>=0.13
xgboost>=2.0
lightgbm>=4.0
statsmodels>=0.14
imbalanced-learn>=0.11
papermill>=2.5
jupyter>=1.0
pyyaml>=6.0
streamlit>=1.30
joblib>=1.3
plotly>=5.18
